requirements.txt
django
djangorestframework
cx_oracle
python-decouple
django-cors-headers
requests